requirements.txt
Django==1.2
PyYAML==3.09
South==0.7.2
BeautifulSoup==3.1.0.1
django-memcached==0.1.2
feedparser==4.1
python-memcached==1.45
simplejson==2.0.9
gunicorn==0.12.0
django-piston
